--- src/components/Mint.svelte
@@ -1,6 +1,0 @@
-<script>
-</script>
-
-<div>
-  <button type="button">Mint</button>
-</div>
